+++ src/lib/proof-system/cache.ts
@@ -7,11 +7,26 @@ import {
 } from '../util/fs.js';
 import { jsEnvironment } from '../../bindings/crypto/bindings/env.js';
 
-export { Cache, CacheHeader, cacheHeaderVersion };
+// external API
+export { Cache, CacheHeader };
+
+// internal API
+export { readCache, writeCache, withVersion, cacheHeaderVersion };
 
 /**
  * Interface for storing and retrieving values, for caching.
  * `read()` and `write()` can just throw errors on failure.
+ *
+ * The data that will be passed to the cache for writing is exhaustively described by the {@link CacheHeader} type.
+ * It represents one of the following:
+ * - The SRS. This is a deterministic lists of curve points (one per curve) that needs to be generated just once,
+ *   to be used for polynomial commitments.
+ * - Lagrange basis commitments. Similar to the SRS, this will be created once for every power-of-2 circuit size.
+ * - Prover and verifier keys for every compiled circuit.
+ *
+ * Per smart contract or ZkProgram, several different keys are created:
+ * - a step prover key (`step-pk`) and verification key (`step-vk`) _for every method_.
+ * - a wrap prover key (`wrap-pk`) and verification key (`wrap-vk`) for the entire contract.
  */
 type Cache = {
   /**
@@ -20,6 +35,7 @@ type Cache = {
    * @param header A small header to identify what is read from the cache.
    */
   read(header: CacheHeader): Uint8Array | undefined;
+
   /**
    * Write a value to the cache.
    *
@@ -27,13 +43,22 @@ type Cache = {
    * @param value The value to write to the cache, as a byte array.
    */
   write(header: CacheHeader, value: Uint8Array): void;
+
   /**
    * Indicates whether the cache is writable.
    */
   canWrite: boolean;
+
+  /**
+   * If `debug` is toggled, `read()` and `write()` errors are logged to the console.
+   *
+   * By default, cache errors are silent, because they don't necessarily represent an error condition,
+   * but could just be a cache miss, or file system permissions incompatible with writing data.
+   */
+  debug?: boolean;
 };
 
-const cacheHeaderVersion = 0.1;
+const cacheHeaderVersion = 1;
 
 type CommonHeader = {
   /**
@@ -54,6 +79,7 @@ type CommonHeader = {
    */
   dataType: 'string' | 'bytes';
 };
+
 type StepKeyHeader<Kind> = {
   kind: Kind;
   programName: string;
@@ -62,9 +88,10 @@ type StepKeyHeader<Kind> = {
   hash: string;
 };
 type WrapKeyHeader<Kind> = { kind: Kind; programName: string; hash: string };
+type PlainHeader<Kind> = { kind: Kind };
 
 /**
- * A header that is passed to the caching layer, to support richer caching strategies.
+ * A header that is passed to the caching layer, to support rich caching strategies.
  *
  * Both `uniqueId` and `programId` can safely be used as a file path.
  */
@@ -73,8 +100,56 @@ type CacheHeader = (
   | StepKeyHeader<'step-vk'>
   | WrapKeyHeader<'wrap-pk'>
   | WrapKeyHeader<'wrap-vk'>
+  | PlainHeader<'srs'>
+  | PlainHeader<'lagrange-basis'>
 ) &
   CommonHeader;
+
+function withVersion(
+  header: Omit<CacheHeader, 'version'>,
+  version = cacheHeaderVersion
+): CacheHeader {
+  let uniqueId = `${header.uniqueId}-${version}`;
+  return { ...header, version, uniqueId } as CacheHeader;
+}
+
+// default methods to interact with a cache
+
+function readCache(cache: Cache, header: CacheHeader): Uint8Array | undefined;
+function readCache<T>(
+  cache: Cache,
+  header: CacheHeader,
+  transform: (x: Uint8Array) => T
+): T | undefined;
+function readCache<T>(
+  cache: Cache,
+  header: CacheHeader,
+  transform?: (x: Uint8Array) => T
+): T | undefined {
+  try {
+    let result = cache.read(header);
+    if (result === undefined) {
+      if (cache.debug) console.trace('cache miss');
+      return undefined;
+    }
+    if (transform === undefined) return result as any as T;
+    return transform(result);
+  } catch (e) {
+    if (cache.debug) console.log('Failed to read cache', e);
+    return undefined;
+  }
+}
+
+function writeCache(cache: Cache, header: CacheHeader, value: Uint8Array) {
+  if (!cache.canWrite) return false;
+  try {
+    cache.write(header, value);
+    return true;
+  } catch (e) {
+    if (cache.debug) console.log('Failed to write cache', e);
+    return false;
+  }
+}
 
 const None: Cache = {
   read() {
@@ -86,7 +161,7 @@ const None: Cache = {
   canWrite: false,
 };
 
-const FileSystem = (cacheDirectory: string): Cache => ({
+const FileSystem = (cacheDirectory: string, debug?: boolean): Cache => ({
   read({ persistentId, uniqueId, dataType }) {
     if (jsEnvironment !== 'node') throw Error('file system not available');
 
@@ -116,6 +191,7 @@ const FileSystem = (cacheDirectory: string): Cache => ({
     });
   },
   canWrite: jsEnvironment === 'node',
+  debug,
 });
 
 const FileSystemDefault = FileSystem(cacheDir('o1js'));
@@ -124,11 +200,17 @@ const Cache = {
   /**
    * Store data on the file system, in a directory of your choice.
    *
+   * Data will be stored in two files per cache entry: a data file and a `.header` file.
+   * The header file just contains a unique string which is used to determine whether we can use the cached data.
+   *
    * Note: this {@link Cache} only caches data in Node.js.
    */
   FileSystem,
   /**
    * Store data on the file system, in a standard cache directory depending on the OS.
+   *
+   * Data will be stored in two files per cache entry: a data file and a `.header` file.
+   * The header file just contains a unique string which is used to determine whether we can use the cached data.
    *
    * Note: this {@link Cache} only caches data in Node.js.
    */