+++ CHANGELOG.md
@@ -17,7 +17,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
  -->
 
-## [Unreleased](https://github.com/o1-labs/o1js/compare/045faa7...HEAD)
+## [Unreleased](https://github.com/o1-labs/o1js/compare/e8e7510e1...HEAD)
+
+> No unreleased changes yet
+
+## [0.14.0](https://github.com/o1-labs/o1js/compare/045faa7...e8e7510e1)
 
 ### Breaking changes